rename & organize e2e utils
