+++ tests/e2e/utils/metrics.py
@@ -2,7 +2,7 @@
 
 import requests
 
-from tests.e2e.constants import BASIC_ENDPOINTS_TIMEOUT
+from tests.e2e.utils.constants import BASIC_ENDPOINTS_TIMEOUT
 
 
 def read_metrics(client):
